o Merge with Linux 2.1.99.  o Fix ancient bug in the ELF loader making ldd crash.  o Fix ancient bug in the keyboard code for SGI, SNI and Jazz.

+++ include/linux/vmalloc.h
@@ -20,7 +20,7 @@ void * vmalloc_uncached(unsigned long size);
 extern void * vmalloc(unsigned long size);
 
 
-int vread(char *buf, char *addr, int count);
+long vread(char *buf, char *addr, unsigned long count);
 void vmfree_area_pages(unsigned long address, unsigned long size);
 int vmalloc_area_pages(unsigned long address, unsigned long size, pgprot_t prot);
